Refactor test infrastructure
