+++ tests/test-utils.h
@@ -1,0 +1,6 @@
+#include "smt.h"
+
+namespace smt_tests {
+
+smt::UnorderedTermSet get_free_symbols(smt::Term & t);
+}
